args: move ArgParseException
