+++ app/src/main/java/checklist/args/def/ArgsDef.java
@@ -1,5 +1,6 @@
 package checklist.args.def;
 
+import checklist.args.ArgParseException;
 import checklist.args.val.ArgsVal;
 import io.vavr.Tuple2;
 import io.vavr.collection.Seq;
