+++ app/src/main/java/checklist/args/def/OptionArgDef.java
@@ -1,5 +1,6 @@
 package checklist.args.def;
 
+import checklist.args.ArgParseException;
 import checklist.args.val.ArgsVal;
 import checklist.args.val.OptionArgVal;
 import io.vavr.Tuple;
